Fixed recommender runtime + black list metodology
