+++ shelter/templates/black_list.html
@@ -22,7 +22,8 @@ th {
   color: white;
 }</style>
 
-<h1 class="display-4 text-center">רשימה שחורה</h1><br>
+<h1 class="display-4 text-center" > רשימה שחורה <i class="fas fa-user-secret"></i></i></h1><br>
+
 <div class="jumbotron">
     <table class="table">
         <thead>
